+++ py/desitarget/cuts.py
@@ -23,6 +23,7 @@ import numpy as np
 import healpy as hp
 from pkg_resources import resource_filename
 import numpy.lib.recfunctions as rfn
+from importlib import import_module
 
 from astropy.table import Table, Row
 
@@ -1844,9 +1845,12 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
         # ADM only northern it will be [False], else it wil be both.
         south_cuts = list(set(np.atleast_1d(photsys_south)))
 
+    # ADM default for target classes we WON'T process is all False.
+    tcfalse = primary & False
+
     # ADM initially set everything to arrays of False for the LRG selection
     # ADM the zeroth element stores the northern targets bits (south=False).
-    lrg_classes = [~primary, ~primary]
+    lrg_classes = [tcfalse, tcfalse]
     if "LRG" in tcnames:
         for south in south_cuts:
             lrg_classes[int(south)] = isLRG(
@@ -1863,7 +1867,7 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
 
     # ADM initially set everything to arrays of False for the ELG selection
     # ADM the zeroth element stores the northern targets bits (south=False).
-    elg_classes = [~primary, ~primary]
+    elg_classes = [tcfalse, tcfalse]
     if "ELG" in tcnames:
         for south in south_cuts:
             elg_classes[int(south)] = isELG(
@@ -1879,7 +1883,7 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
 
     # ADM initially set everything to arrays of False for the QSO selection
     # ADM the zeroth element stores the northern targets bits (south=False).
-    qso_classes = [[~primary, ~primary], [~primary, ~primary]]
+    qso_classes = [[tcfalse, tcfalse], [tcfalse, tcfalse]]
     if "QSO" in tcnames:
         for south in south_cuts:
             if qso_selection == 'colorcuts':
@@ -1915,7 +1919,7 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
 
     # ADM initially set everything to arrays of False for the BGS selection
     # ADM the zeroth element stores the northern targets bits (south=False).
-    bgs_classes = [[~primary, ~primary, ~primary], [~primary, ~primary, ~primary]]
+    bgs_classes = [[tcfalse, tcfalse, tcfalse], [tcfalse, tcfalse, tcfalse]]
     # ADM set the BGS bits
     if "BGS" in tcnames:
         for south in south_cuts:
@@ -1955,8 +1959,8 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
 
     # ADM initially set everything to arrays of False for the MWS selection
     # ADM the zeroth element stores the northern targets bits (south=False).
-    mws_classes = [[~primary, ~primary, ~primary], [~primary, ~primary, ~primary]]
-    mws_nearby = ~primary
+    mws_classes = [[tcfalse, tcfalse, tcfalse], [tcfalse, tcfalse, tcfalse]]
+    mws_nearby = tcfalse
     if "MWS" in tcnames:
         mws_nearby = isMWS_nearby(
             gaia=gaia, gaiagmag=gaiagmag, parallax=parallax,
@@ -1977,7 +1981,7 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
 
     # ADM treat the MWS WD selection specially, as we have to run the
     # ADM white dwarfs for standards and MWS science targets.
-    mws_wd = ~primary
+    mws_wd = tcfalse
     if "MWS" in tcnames or "STD" in tcnames:
         mws_wd = isMWS_WD(
             gaia=gaia, galb=galb, astrometricexcessnoise=gaiaaen,
@@ -1988,7 +1992,7 @@ def set_target_bits(photsys_north, photsys_south, obs_rflux,
         )
 
     # ADM initially set everything to False for the standards.
-    std_faint, std_bright, std_wd = ~primary, ~primary, ~primary
+    std_faint, std_bright, std_wd = tcfalse, tcfalse, tcfalse
     if "STD" in tcnames:
         # ADM run the MWS_MAIN target types for both faint and bright.
         # ADM Make sure to pass all of the needed columns! At one point we stopped
@@ -2118,19 +2122,27 @@ def apply_cuts_gaia(numproc=4, survey='main', nside=None, pixlist=None):
         - Only run on Gaia-only target selections.
         - The environment variable $GAIA_DIR must be set.
 
-    See desitarget.sv1.sv1_targetmask.desi_mask
-    and desitarget.targetmask.desi_mask for bit definitions.
+    See desitarget.svX.svX_targetmask.desi_mask or
+    desitarget.targetmask.desi_mask for bit definitions.
     """
     # ADM set different bits based on whether we're using the main survey
     # code or an iteration of SV.
     if survey == 'main':
         import desitarget.cuts as targcuts
         from desitarget.targetmask import desi_mask, mws_mask
-    elif survey == 'sv1':
-        import desitarget.sv1.sv1_cuts as targcuts
-        from desitarget.sv1.sv1_targetmask import desi_mask, mws_mask
+    elif survey[:2] == 'sv':
+        try:
+            targcuts = import_module("desitarget.{}.{}_cuts".format(survey, survey))
+            targmask = import_module("desitarget.{}.{}_targetmask".format(
+                survey, survey))
+        except ModuleNotFoundError:
+            msg = 'Bitmask yaml or cuts do not exist for survey type {}'.format(
+                survey)
+            log.critical(msg)
+            raise ModuleNotFoundError(msg)
+        desi_mask, mws_mask = targmask.desi_mask, targmask.mws_mask
     else:
-        msg = "survey must be either 'main'or 'sv1', not {}!!!".format(survey)
+        msg = "survey must be either 'main'or 'svX', not {}!!!".format(survey)
         log.critical(msg)
         raise ValueError(msg)
 
@@ -2280,10 +2292,10 @@ def apply_cuts(objects, qso_selection='randomforest', gaiamatch=False,
     # code or an iteration of SV.
     if survey == 'main':
         import desitarget.cuts as targcuts
-    elif survey == 'sv1':
-        import desitarget.sv1.sv1_cuts as targcuts
+    elif survey[:2] == 'sv':
+        targcuts = import_module("desitarget.{}.{}_cuts".format(survey, survey))
     else:
-        msg = "survey must be either 'main'or 'sv1', not {}!!!".format(survey)
+        msg = "survey must be either 'main'or 'svX', not {}!!!".format(survey)
         log.critical(msg)
         raise ValueError(msg)